scripts/scrapers/requirements.txt
# EUDAMED & NMPA 爬虫依赖

# 浏览器自动化
playwright>=1.40.0

# 数据库
supabase>=2.0.0

# 工具
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0
